Simplify module declaration
A recurring pattern in hug is to declare a module as an empty table,
then declare a metatable for that module as `mt`. This commit pushes
that implementation detail into a function so that module declaration
isn't so ugly. From now on, all modules will be expected to contain
metamethods directly as fields so that they can be easily accessed and
reused.